Refactor ItemSearchDTO to own display block-name resolution
- Add ItemSearchDTO#getDisplayBlockName() to encapsulate block-name translation/fallback logic alongside getDisplayItemName().
- Move block display-name translation from StoragePosHighlighter into ItemSearchDTO.
- Simplify StoragePosHighlighter by using DTO display accessors for both block and item names.
- Remove duplicated translateBlockName logic from highlighter.
- Keep behavior unchanged while improving cohesion and reducing UI-side duplication.
